+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.4.9
 - vs 18.6.2
+- add missing configureawait(false)
 
 ## 2.4.8
 - vs 18.6.0